+++ NEWS.md
@@ -1,3 +1,15 @@
+# spfilteR 1.1.0
+
+- CRAN resubmission
+- fix minor bug when checking 'tol' in `lmFilter()` function
+- new vignette name
+- add new functions
+    - `MI.local()` function to calculate local Moran's I
+    - `vp()` function for variation partitioning
+- update reference to `MI.local()` in documentation files
+
+---
+
 # spfilteR 1.0.0.9000
 
 - fix minor bug when checking 'tol' in `lmFilter()` function